ner/bert-crf-requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Keras==2.2.4
Keras-Applications==1.0.8
keras-bert==0.80.0
keras-contrib==2.0.8
keras-embed-sim==0.8.0
keras-layer-normalization==0.14.0
keras-multi-head==0.27.0
keras-pos-embd==0.11.0
keras-position-wise-feed-forward==0.6.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.46.0
keras-transformer==0.31.0
Markdown==3.2.2
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.48.2
traitlets==4.3.3
typeguard==2.9.1
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
